@@ -6,7 +6,7 @@ import {
   HttpEvent
 } from '@angular/common/http';
 
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { CookieService } from 'ngx-cookie';
 
 @Injectable()
