@@ -1,8 +1,10 @@
-import { StaticInjector } from '@angular/core/src/di/injector';
-import { ModalService } from './modal.service';
-import { ComponentRef, Output, Input, Injector, EventEmitter } from '@angular/core';
-import { element } from 'protractor';
 
+import { ModalService } from './modal.service';
+import { Output, Input, EventEmitter, Directive } from '@angular/core';
+
+@Directive({
+    selector: '[modal]'
+  })
 export abstract class Modal {
     protected name: string;
     protected x: number;
